@@ -18,26 +18,29 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=3.4.0-282.3.beta <4.0.0'
 
 dependencies:
-  animated_splash_screen: ^1.0.0+1
+  animated_splash_screen: ^1.3.0
   flutter:
     sdk: flutter
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.8
+  page_transition: ^2.1.0
 
 dev_dependencies:
-  flutter_launcher_name: ^0.0.1
-  flappy_search_bar: ^1.7.2
+  launcher_name: ^1.0.2
+  flappy_search_bar_ns: ^2.0.2
   flutter_test:
     sdk: flutter
 
-flutter_launcher_name:
-  name: "Pakaian Adat App"
+  flutter_lints: ^4.0.0
+  
+launcher_name:
+  default: "Pakaian Adat App"
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
