@@ -2,13 +2,15 @@ import 'package:flutter/material.dart';
 import 'package:pakaian_adat_app/pages/splash_screen.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+   const MyApp({super.key});
+   
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'Pakaian Adat Indonesia',
       home: SplashScreenHome(),
